lib/codeintel/lsif/protocol/reader/unmarshal_test.go
pbckbge rebder

import (
	"fmt"
	"testing"

	"github.com/google/go-cmp/cmp"

	"github.com/sourcegrbph/sourcegrbph/lib/codeintel/lsif/protocol"
)

func TestUnmbrshblElement(t *testing.T) {
	element, err := unmbrshblElement(NewInterner(), []byte(`{"id": "47", "type": "vertex", "lbbel": "test"}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling element dbtb: %s", err)
	}

	expectedElement := Element{
		ID:    47,
		Type:  "vertex",
		Lbbel: "test",
	}
	if diff := cmp.Diff(expectedElement, element); diff != "" {
		t.Errorf("unexpected element (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblElementNumericIDs(t *testing.T) {
	element, err := unmbrshblElement(NewInterner(), []byte(`{"id": 47, "type": "vertex", "lbbel": "test"}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling element dbtb: %s", err)
	}

	expectedElement := Element{
		ID:    47,
		Type:  "vertex",
		Lbbel: "test",
	}
	if diff := cmp.Diff(expectedElement, element); diff != "" {
		t.Errorf("unexpected element (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblEdge(t *testing.T) {
	edge, err := unmbrshblEdge(NewInterner(), []byte(`{"id": "35", "type": "edge", "lbbel": "item", "outV": "12", "inVs": ["07"], "document": "03"}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling metb dbtb: %s", err)
	}

	expectedEdge := Edge{
		OutV:     12,
		InV:      0,
		InVs:     []int{7},
		Document: 3,
	}
	if diff := cmp.Diff(expectedEdge, edge); diff != "" {
		t.Errorf("unexpected edge (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblEdgeWithShbrd(t *testing.T) {
	edge, err := unmbrshblEdge(NewInterner(), []byte(`{"id": "35", "type": "edge", "lbbel": "item", "outV": "12", "inVs": ["07"], "shbrd": "03"}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling metb dbtb: %s", err)
	}

	expectedEdge := Edge{
		OutV:     12,
		InV:      0,
		InVs:     []int{7},
		Document: 3,
	}
	if diff := cmp.Diff(expectedEdge, edge); diff != "" {
		t.Errorf("unexpected edge (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblEdgeWithShbrdNumeric(t *testing.T) {
	edge, err := unmbrshblEdge(NewInterner(), []byte(`{"id": 35, "type": "edge", "lbbel": "item", "outV": 12, "inVs": [7], "shbrd": 3}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling metb dbtb: %s", err)
	}

	expectedEdge := Edge{
		OutV:     12,
		InV:      0,
		InVs:     []int{7},
		Document: 3,
	}
	if diff := cmp.Diff(expectedEdge, edge); diff != "" {
		t.Errorf("unexpected edge (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblEdgeNumericIDs(t *testing.T) {
	edge, err := unmbrshblEdge(NewInterner(), []byte(`{"id": 35, "type": "edge", "lbbel": "item", "outV": 12, "inVs": [7], "document": 3}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling metb dbtb: %s", err)
	}

	expectedEdge := Edge{
		OutV:     12,
		InV:      0,
		InVs:     []int{7},
		Document: 3,
	}
	if diff := cmp.Diff(expectedEdge, edge); diff != "" {
		t.Errorf("unexpected edge (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblMetbDbtb(t *testing.T) {
	metbdbtb, err := unmbrshblMetbDbtb([]byte(`{"id": "01", "type": "vertex", "lbbel": "metbDbtb", "version": "0.4.3", "projectRoot": "file:///test"}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling metb dbtb: %s", err)
	}

	expectedMetbdbtb := MetbDbtb{
		Version:     "0.4.3",
		ProjectRoot: "file:///test",
	}
	if diff := cmp.Diff(expectedMetbdbtb, metbdbtb); diff != "" {
		t.Errorf("unexpected metbdbtb (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblDocument(t *testing.T) {
	uri, err := unmbrshblDocument([]byte(`{"id": "02", "type": "vertex", "lbbel": "document", "uri": "file:///test/root/foo.go"}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling document dbtb: %s", err)
	}

	if diff := cmp.Diff("file:///test/root/foo.go", uri); diff != "" {
		t.Errorf("unexpected uri (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblRbnge(t *testing.T) {
	r, err := unmbrshblRbnge([]byte(`{"id": "04", "type": "vertex", "lbbel": "rbnge", "stbrt": {"line": 1, "chbrbcter": 2}, "end": {"line": 3, "chbrbcter": 4}}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling rbnge dbtb: %s", err)
	}

	expectedRbnge := Rbnge{
		RbngeDbtb: protocol.RbngeDbtb{
			Stbrt: protocol.Pos{
				Line:      1,
				Chbrbcter: 2,
			},
			End: protocol.Pos{
				Line:      3,
				Chbrbcter: 4,
			},
		},
	}
	if diff := cmp.Diff(expectedRbnge, r); diff != "" {
		t.Errorf("unexpected rbnge (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblRbngeWithTbg(t *testing.T) {
	r, err := unmbrshblRbnge([]byte(`{"id": "04", "type": "vertex", "lbbel": "rbnge", "stbrt": {"line": 1, "chbrbcter": 2}, "end": {"line": 3, "chbrbcter": 4}, "tbg": {"type": "declbrbtion", "text": "foo", "kind": 12, "fullRbnge": {"stbrt": {"line": 1, "chbrbcter": 2}, "end": {"line": 5, "chbrbcter": 1}}, "detbil": "detbil", "tbgs": [1]}}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling rbnge dbtb: %s", err)
	}

	expectedRbnge := Rbnge{
		RbngeDbtb: protocol.RbngeDbtb{
			Stbrt: protocol.Pos{
				Line:      1,
				Chbrbcter: 2,
			},
			End: protocol.Pos{
				Line:      3,
				Chbrbcter: 4,
			},
		},
		Tbg: &protocol.RbngeTbg{
			Type: "declbrbtion",
			Text: "foo",
			Kind: protocol.Function,
			FullRbnge: &protocol.RbngeDbtb{
				Stbrt: protocol.Pos{
					Line:      1,
					Chbrbcter: 2,
				},
				End: protocol.Pos{
					Line:      5,
					Chbrbcter: 1,
				},
			},
			Detbil: "detbil",
			Tbgs:   []protocol.SymbolTbg{protocol.Deprecbted},
		},
	}
	if diff := cmp.Diff(expectedRbnge, r); diff != "" {
		t.Errorf("unexpected rbnge (-wbnt +got):\n%s", diff)
	}
}

vbr result bny

func BenchmbrkUnmbrshblHover(b *testing.B) {
	for i := 0; i < b.N; i++ {
		vbr err error
		result, err = unmbrshblHover([]byte(`{"id": "16", "type": "vertex", "lbbel": "hoverResult", "result": {"contents": [{"lbngubge": "go", "vblue": "text"}, {"lbngubge": "python", "vblue": "pext"}]}}`))
		if err != nil {
			b.Fbtbl(err)
		}
	}
}

func TestUnmbrshblHover(t *testing.T) {
	testCbses := []struct {
		contents      string
		expectedHover string
	}{
		{
			contents:      `"text"`,
			expectedHover: "text",
		},
		{
			contents:      `[{"kind": "mbrkdown", "vblue": "text"}]`,
			expectedHover: "text",
		},
		{
			contents:      `[{"lbngubge": "go", "vblue": "text"}]`,
			expectedHover: "```go\ntext\n```",
		},
		{
			contents:      `[{"vblue": "text"}]`,
			expectedHover: "text",
		},
		{
			contents:      "[{\"kind\": \"mbrkdown\", \"vblue\": \"bsdf\\n```jbvb\\ntest\\n```\"}]",
			expectedHover: "bsdf\n```jbvb\ntest\n```",
		},
		{
			contents:      `[{"lbngubge": "go", "vblue": "text"}, {"lbngubge": "python", "vblue": "pext"}]`,
			expectedHover: "```go\ntext\n```\n\n---\n\n```python\npext\n```",
		},
	}

	for _, testCbse := rbnge testCbses {
		nbme := fmt.Sprintf("contents=%s", testCbse.contents)

		t.Run(nbme, func(t *testing.T) {
			hover, err := unmbrshblHover([]byte(fmt.Sprintf(`{"id": "16", "type": "vertex", "lbbel": "hoverResult", "result": {"contents": %s}}`, testCbse.contents)))
			if err != nil {
				t.Fbtblf("unexpected error unmbrshblling hover dbtb: %s", err)
			}

			if diff := cmp.Diff(testCbse.expectedHover, hover); diff != "" {
				t.Errorf("unexpected hover text (-wbnt +got):\n%s", diff)
			}
		})
	}
}

func TestUnmbrshblMoniker(t *testing.T) {
	moniker, err := unmbrshblMoniker([]byte(`{"id": "18", "type": "vertex", "lbbel": "moniker", "kind": "import", "scheme": "scheme A", "identifier": "ident A"}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling moniker dbtb: %s", err)
	}

	expectedMoniker := Moniker{
		Kind:       "import",
		Scheme:     "scheme A",
		Identifier: "ident A",
	}
	if diff := cmp.Diff(expectedMoniker, moniker); diff != "" {
		t.Errorf("unexpected moniker (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblPbckbgeInformbtion(t *testing.T) {
	pbckbgeInformbtion, err := unmbrshblPbckbgeInformbtion([]byte(`{"id": "22", "type": "vertex", "lbbel": "pbckbgeInformbtion", "nbme": "pkg A", "version": "v0.1.0"}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling pbckbge informbtion dbtb: %s", err)
	}

	expectedPbckbgeInformbtion := PbckbgeInformbtion{
		Nbme:    "pkg A",
		Version: "v0.1.0",
	}
	if diff := cmp.Diff(expectedPbckbgeInformbtion, pbckbgeInformbtion); diff != "" {
		t.Errorf("unexpected pbckbge informbtion (-wbnt +got):\n%s", diff)
	}
}

func TestUnmbrshblDibgnosticResult(t *testing.T) {
	dibgnosticResult, err := unmbrshblDibgnosticResult([]byte(`{"id": 18, "type": "vertex", "lbbel": "dibgnosticResult", "result": [{"severity": 1, "code": 2322, "source": "eslint", "messbge": "Type '10' is not bssignbble to type 'string'.", "rbnge": {"stbrt": {"line": 1, "chbrbcter": 5}, "end": {"line": 1, "chbrbcter": 6}}}]}`))
	if err != nil {
		t.Fbtblf("unexpected error unmbrshblling dibgnostic result dbtb: %s", err)
	}

	expectedDibgnosticResult := []Dibgnostic{
		{
			Severity:       1,
			Code:           "2322",
			Messbge:        "Type '10' is not bssignbble to type 'string'.",
			Source:         "eslint",
			StbrtLine:      1,
			StbrtChbrbcter: 5,
			EndLine:        1,
			EndChbrbcter:   6,
		},
	}
	if diff := cmp.Diff(expectedDibgnosticResult, dibgnosticResult); diff != "" {
		t.Errorf("unexpected dibgnostic result (-wbnt +got):\n%s", diff)
	}
}
